replaced float with double

+++ ActivationFunction_I.java
@@ -4,5 +4,5 @@
 @FunctionalInterface
 interface ActivationFunction_I {
 
-    float apply(float input);
+    double apply(double input);
 }
